Add feature to add new user to a mailing group using notification message.

+++ handlers/admin_handlers.py
@@ -2,15 +2,17 @@ from aiogram import Router, F, Bot
 from aiogram.fsm.context import FSMContext
 from aiogram.types import Message, CallbackQuery
 
+from config_data import DATETIME_FORMAT
 from filters.role_filters import RoleFilter
-from project_types.enum_types import ChatRole, NoticeAction
-from project_types.bot_config import BotConfig
+from fsms import NewUserStates
 from keboards import admin_kb as kb
-from keboards.common_kb import HOME_KB
+from keboards.common_kb import HOME_KB, ADD_TO_GROUP_HOME_KB, EDIT_GROUPS_HOME, edit_groups_kb
 from keboards.inline_kb import NewUserNotice
-from texts.common_texts import HOME, APPLICANT_NOTICE_APPROVED, APPLICANT_NOTICE_DECLINED, APPLICANT_NOTICE_POSTPONED
+from project_types.bot_config import BotConfig
+from project_types.enum_types import ChatRole, NoticeAction
+from texts.common_texts import (HOME, APPLICANT_NOTICE_DECLINED, APPLICANT_NOTICE_POSTPONED, APPLICANT_APPROVED,
+                                applicant_declined_text, applicant_approved_text)
 from texts.new_user import APPLICANT_APPROVED, APPLICANT_DECLINED
-
 
 router = Router()
 
@@ -22,22 +24,62 @@ async def home(message: Message, config: BotConfig, state: FSMContext):
 
 
 @router.callback_query(NewUserNotice.filter())
-async def new_app_notice(callback_query: CallbackQuery, callback_data:NewUserNotice, config: BotConfig, bot: Bot):
-    if callback_data.action == NoticeAction.APPROVE:
-        await config.change_subscriptions(user_id=callback_data.user_id)
-        await config.alter_user_role(user_id=callback_data.user_id, new_role=ChatRole.USER)
-        await bot.send_message(chat_id=callback_data.user_id, text=APPLICANT_APPROVED, reply_markup=HOME_KB)
-        await callback_query.answer(text=APPLICANT_NOTICE_APPROVED)
-    if callback_data.action == NoticeAction.DECLINE:
-        c_info = await config.get_user_info(user_id=callback_data.user_id)
-        if c_info is not None:
-            await config.remove_user_info(user_id=callback_data.user_id)
-        await config.remove_user_by_id(user_id=callback_data.user_id)
-        await bot.send_message(chat_id=callback_data.user_id, text=APPLICANT_DECLINED)
-        await callback_query.answer(text=APPLICANT_NOTICE_DECLINED)
-    if callback_data.action == NoticeAction.POSTPONE:
-        await callback_query.answer(text=callback_query.message.text)
-        await callback_query.message.delete()
-        await callback_query.answer(text=APPLICANT_NOTICE_POSTPONED)
+async def new_app_notice(callback_query: CallbackQuery, callback_data: NewUserNotice,
+                         config: BotConfig, bot: Bot, state: FSMContext):
+    user = config.users.get(callback_data.user_id)
+    admin = config.users.get(callback_query.from_user.id)
+    m_app = await config.get_application(applicant_id=callback_data.user_id)
+    if user is not None and user.role == ChatRole.APPLICANT:
+        if callback_data.action == NoticeAction.APPROVE:
+            # Add new user to the bot
+            await config.change_subscriptions(user_id=user.id)
+            await config.alter_user_role(user_id=user.id, new_role=ChatRole.USER)
+            await config.close_application(applicant_id=user.id, admin_id=admin.id)
+            await bot.send_message(chat_id=user.id, text=APPLICANT_APPROVED, reply_markup=HOME_KB)
+
+            # Offer admin to add new user to mailing groups
+            await bot.send_message(chat_id=admin.id,
+                                   text= APPLICANT_APPROVED,
+                                   reply_markup=ADD_TO_GROUP_HOME_KB)
+            await state.set_state(NewUserStates.add_to_group)
+            await state.update_data(user_id=user.id)
+        if callback_data.action == NoticeAction.DECLINE:
+            await config.remove_user_by_id(user_id=user.id)
+            await config.close_application(applicant_id=user.id, admin_id=admin.id)
+            await bot.send_message(chat_id=user.id, text=APPLICANT_DECLINED)
+            await callback_query.answer(text=APPLICANT_NOTICE_DECLINED)
+        if callback_data.action == NoticeAction.POSTPONE:
+            await callback_query.answer(text=callback_query.message.text)
+            await callback_query.message.delete()
+            await callback_query.answer(text=APPLICANT_NOTICE_POSTPONED)
+    elif user is None:
+        await callback_query.answer(text=applicant_declined_text(amin_name=admin.full_name(),
+                                                                date=m_app.decision_date.strftime(DATETIME_FORMAT)))
+    else:
+        await callback_query.answer(text=applicant_approved_text(amin_name=admin.full_name(),
+                                                                date=m_app.decision_date.strftime(DATETIME_FORMAT)))
+
+
+@router.message(F.text == 'Добавить в группу', NewUserStates.add_to_group)
+@router.message(F.text == 'Редактировать группы пользователя', NewUserStates.add_to_group)
+async def chose_group(message: Message, config: BotConfig, state: FSMContext):
+    data = await state.get_data()
+    user = config.users[data['user_id']]
+    await message.answer(text='Выберите группу',
+                         reply_markup=edit_groups_kb(user=user, all_groups=list(config.all_groups.keys()), back=False))
+    await state.set_state(NewUserStates.chose_group)
+
+
+@router.message(F.text != 'Домой', NewUserStates.chose_group)
+async def add_user_to_group(message: Message, config: BotConfig, state: FSMContext):
+    data = await state.get_data()
+    user = config.users[data['user_id']]
+    group = message.text.split(' ')[-1]
+    await config.add_user_to_mailing_group(user.id, group)
+    await message.answer(text=user.represent_user_full())
+    await message.answer(text=f'Пользователь {user.full_name()} успешно добавлен в группу {group}',
+                         reply_markup=EDIT_GROUPS_HOME)
+    await state.set_state(NewUserStates.add_to_group)
+
